requirements.txt
Django==4.0.2
Pillow==9.0.1
PyYAML==6.0
black==22.1.0
dj-database-url==0.5.0
dj-rest-auth==2.2.3
django-allauth==0.48.0
django-cors-headers==3.11.0
djangorestframework==3.13.1
factory-boy==3.2.1
gunicorn==20.1.0
python-dotenv==0.19.2
uritemplate==4.1.1
whitenoise==6.0.0
